pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zerobus-station"
version = "1.0.0"
description = "Configuration-driven FastAPI service for ingesting data into Databricks via Zerobus"
requires-python = ">=3.11"
keywords = ["zerobus", "databricks", "fastapi", "ingestion"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-dotenv>=1.0.0",
    "databricks-zerobus-ingest-sdk>=0.2.0",
    "grpcio-tools>=1.76.0",
]
